// src/app/models/scenic-spot.model.ts
// Generated by https://quicktype.io

export interface ScenicSpotModel {
  ID: string;
  Name: string;
  DescriptionDetail: string;
  Phone: string;
  ZipCode: string;
  OpenTime?: string;
  Picture: Picture;
  Position: Position;
  Class1?: string;
  Level: string;
  WebsiteUrl?: string;
  ParkingPosition: ParkingPosition;
  City: string;
  SrcUpdateTime: Date;
  UpdateTime: Date;
  Class2?: string;
  Class3?: string;
}



export interface ParkingPosition {
}

export interface Picture {
  PictureUrl1: string;
  PictureDescription1: string;
}

export interface Position {
  PositionLon: number;
  PositionLat: number;
  GeoHash: string;
}